Module Bos_setup

Quick setup for simple programs.

Linking against this module setups Logs and issuing:

open Bos_setup

in a module is sufficient to bring Rresult, Astring and Bos in scope. See also how to use this for interpreted programs.

v0.2.0 - homepage

Interpreted programs

Results

type ('a, 'b) result = ('a, 'b) Rresult.result =

| Ok of 'a

| Error of 'b

The type for results.

val (>>=) : ('a, 'b) Result.result -> ('a -> ('c, 'b) Result.result) -> ('c, 'b) Result.result

(>>=) is R.( >>= ).

val (>>|) : ('a, 'b) Result.result -> ('a -> 'c) -> ('c, 'b) Result.result

(>>|) is R.( >>| ).

module R : sig ... end

Astring

val strf : ('a, Stdlib.Format.formatter, unit, string) Stdlib.Pervasives.format4 -> 'a

strf is Astring.strf.

val (^) : string -> string -> string

^ is Astring.(^).

module Char : sig ... end

module String : sig ... end

Bos

module Pat : sig ... end

module Cmd : sig ... end

module OS : sig ... end

Fmt & Logs

module Fmt : sig ... end

module Logs : sig ... end