Module Bos

Basic OS interaction.

Open the module to use it, this defines only modules in your scope.

WARNING. This API is still subject to change in the future but feedback and suggestions are welcome on the project's issue tracker.

Basic types

module Pat : sig ... end

Named string patterns.

module Cmd : sig ... end

Command lines.

OS interaction

module OS : sig ... end

OS interaction