Module Bos.OS

OS interaction

Results

The functions of this module never raise Sys_error or Unix.Unix_error instead they turn these errors into error messages. If you need fine grained control over unix errors use the lower level functions in Bos.OS.U.

type ('a, 'e) result = ('a, [> Rresult.R.msg ] as 'e) result

The type for OS results.

Environment variables and program arguments

module Env : sig ... end

Environment variables.

module Arg : sig ... end

Quick and dirty program arguments parsing.

File system operations

Note. When paths are relative they are expressed relative to the current working directory.

module Path : sig ... end

Path operations.

module File : sig ... end

File operations.

module Dir : sig ... end

Directory operations.

Commands

module Cmd : sig ... end

Executing commands.

Low level access

module U : sig ... end

Low level Unix access.